+++ libs/data-access/crypto/src/index.ts
@@ -1,4 +1,0 @@
-export * from './lib/crypto.service';
-export * from './lib/verification.service';
-export * from './lib/devices.service';
-export * from './lib/crypto-spike.service';